pyproject.toml
[project]
name = "x11-window-interactor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cython>=3.0.12",
    "numpy>=2.2.4",
    "opencv-python>=4.11.0.86",
    "python-xlib>=0.33",
]
